including sdk files
As long as Schoology has not been modified, and the changes are not in any branch on github, I will keep the copy of the API right here (although I know that's not recommended). API development is very quiet on the Schoology front, I don't think we'll have issues. This commit is just for a couple of (crucial) lines of code. See my recorded issue in github