New version (#13)
* new release

almost everything changed. Especially new folder for classes. Code split into controller and view.

* use local icon

better than hotlinking, but in fact there should be a custom logo. Thinking of just a reversed color schoology logo,

* readme should have changed on master

* including sdk files

As long as Schoology has not been modified, and the changes are not in any branch on github, I will keep the copy of the API right here (although I know that's not recommended). API development is very quiet on the Schoology front, I don't think we'll have issues. This commit is just for a couple of (crucial) lines of code. See my recorded issue in github

* Entire new version

Changes:
- Cache: using SQLite database to cache api results, and minimize time taken for each request
- Course: features like pictures, list, clickable user
- Filesystem: all FS related functions, like mkdir, delete, etc
- Html: generate HTML output like select, link etc
- Submission: special class for the many submission related fields
- Timer: make sure no more than 50 api requests per second go out
- Uploader: upload a csv with members (in beta status)
- groups.php: better navigation

* explain new version

* File not required anymore

closes #6

* $formatId should be plain formatId

closes #5

* fix filter, show timestamp

closes #8, #10

* fix filter, and use large size pictures

closes #11, closes #8

* default submission status ALL

closes #12