closes #52 drag and drop upload files for import
This is half of the implementation. The actual creation of groups is not implemented yet. That still needs some refactoring, such as a check for enrollment of a student.
Also in this solve the UTF-8 of the export (see #27) function has hopefully been fixed now. This has been reverted a couple of times. Perhaps the data from the api has changed?