closes #59 regular update of sqlite database records
This is implemented in front end (refresh button) and also using a cron task (on server labs.bluekoala-aruba.com). The number of records to update can be modified using url parameter 'n' (defaults to 1). Ajax option is not implemented. The cahce has a TLRU (time-aware least recently used) system, but in this case we simply fetch the oldest record.